- removed a lot of queries as they will not be needed - removed Relation in favor of a Link (link is a kind of relation now)
