--- src/Edde/Api/Query/IUpdateRelationQuery.php
@@ -1,5 +1,0 @@
-<?php
-	namespace Edde\Api\Query;
-
-		interface IUpdateRelationQuery extends ICreateRelationQuery {
-		}
