+++ src/Edde/Api/Schema/ITarget.php
@@ -1,0 +1,21 @@
+<?php
+	namespace Edde\Api\Schema;
+
+	/**
+	 * Target of a link.
+	 */
+		interface ITarget {
+			/**
+			 * get schema of a target
+			 *
+			 * @return ISchema
+			 */
+			public function getSchema(): ISchema;
+
+			/**
+			 * get target property
+			 *
+			 * @return IProperty
+			 */
+			public function getProperty(): IProperty;
+		}
